Refactor and add to centroids testing
